Split up accountsjs package for better structure
