+++ packages/core-accountsjs/accounts-server.js
@@ -1,42 +1,10 @@
 import { AccountsServer, ServerHooks } from '@accounts/server';
-import { AccountsPassword } from '@accounts/password';
-import MongoDBInterface from '@accounts/mongo';
-import { MongoInternals } from 'meteor/mongo';
-import { DatabaseManager } from '@accounts/database-manager';
-import crypto, { randomBytes } from 'crypto';
 import { WorkerDirector } from 'meteor/unchained:core-worker';
-
-export const randomValueHex = (len) => {
-  return crypto
-    .randomBytes(Math.ceil(len / 2))
-    .toString('hex') // convert to hexadecimal format
-    .slice(0, len); // return required number of characters
-};
-
-const METEOR_ID_LENGTH = 17;
-
-const idProvider = () =>
-  randomBytes(30)
-    .toString('base64')
-    .replace(/[\W_]+/g, '')
-    .substr(0, METEOR_ID_LENGTH);
-
-const dateProvider = (date) => date || new Date();
-
-const mongoStorage = new MongoDBInterface(
-  MongoInternals.defaultRemoteCollectionDriver().mongo.db,
-  {
-    convertUserIdToMongoObjectId: false,
-    convertSessionIdToMongoObjectId: false,
-    idProvider,
-    dateProvider,
-  }
-);
-
-export const dbManager = new DatabaseManager({
-  sessionStorage: mongoStorage,
-  userStorage: mongoStorage,
-});
+import crypto from 'crypto';
+import { randomValueHex } from './helpers';
+import settings from './settings';
+import { dbManager } from './db-manager';
+import { accountsPassword } from './accounts-password';
 
 const accountsServerOptions = {
   siteUrl: process.env.ROOT_URL,
@@ -74,16 +42,10 @@ const accountsServerOptions = {
   },
 };
 
-class UnchainedAccountsPassword extends AccountsPassword {}
-
-export const accountsPassword = new UnchainedAccountsPassword({
-  sendVerificationEmailAfterSignup: true,
-});
-
 class UnchainedAccountsServer extends AccountsServer {
-  DEFAULT_LOGIN_EXPIRATION_DAYS = 90;
+  DEFAULT_LOGIN_EXPIRATION_DAYS = settings.DEFAULT_LOGIN_EXPIRATION_DAYS;
 
-  LOGIN_UNEXPIRING_TOKEN_DAYS = 365 * 100;
+  LOGIN_UNEXPIRING_TOKEN_DAYS = settings.LOGIN_UNEXPIRING_TOKEN_DAYS;
 
   destroyToken = (userId, loginToken) => {
     this.users.update(userId, {
@@ -164,6 +126,7 @@ class UnchainedAccountsServer extends AccountsServer {
   }
 }
 
+// eslint-disable-next-line import/prefer-default-export
 export const accountsServer = new UnchainedAccountsServer(
   {
     db: dbManager,
