+++ packages/core-accountsjs/accountsjs.js
@@ -1,0 +1,8 @@
+import settings from './settings';
+
+export * from './helpers';
+export * from './accounts-server';
+export * from './accounts-password';
+export default (options) => {
+  settings.load(options);
+};
