--- packages/core-accountsjs/core-accountsjs-tests.js
@@ -1,1 +1,0 @@
-// TBH
