+++ packages/core-accountsjs/helpers.js
@@ -1,0 +1,18 @@
+import crypto, { randomBytes } from 'crypto';
+
+export const randomValueHex = (len) => {
+  return crypto
+    .randomBytes(Math.ceil(len / 2))
+    .toString('hex') // convert to hexadecimal format
+    .slice(0, len); // return required number of characters
+};
+
+const METEOR_ID_LENGTH = 17;
+
+export const idProvider = () =>
+  randomBytes(30)
+    .toString('base64')
+    .replace(/[\W_]+/g, '')
+    .substr(0, METEOR_ID_LENGTH);
+
+export const dateProvider = (date) => date || new Date();
